+++ src/lib/demo.svelte
@@ -2,6 +2,7 @@
 	/**
 	 * import namor for random data generation
 	 */
+	export let dataSet;
 	import namor from 'namor';
 	import SvelTable from 'sveltable';
 
@@ -9,26 +10,7 @@
 	 * newPerson function
 	 * @param num = number of rows to create
 	 */
-	function newPerson(num) {
-		const sexAssigner = Math.random();
-		return {
-			id: num,
-			firstName: namor.generate({ words: 1, numbers: 0 }),
-			lastName: namor.generate({ words: 1, numbers: 0 }),
-			sex: sexAssigner > 0.75 ? 'prefer not to say' : sexAssigner > 0.25 ? 'male' : 'female',
-			age: Math.floor(Math.random() * 110),
-			occupation: namor.generate({ words: 1, numbers: 0 })
-		};
-	}
-
-	let userDefinedNum = 1000;
-
-	let data = [];
-
-	for (let i = 1; i < userDefinedNum + 1; i++) {
-		let newP = newPerson(i);
-		data.push(newP);
-	}
+	
 </script>
 
 <div class="SectionContainer">
@@ -37,7 +19,7 @@
 	</div>
 
 	<div class="DemoContainer">
-		<SvelTable dataSet={data} />
+		<SvelTable dataSet={dataSet} />
 	</div>
 </div>
