+++ src/routes/index.svelte
@@ -15,6 +15,18 @@
 	import Demo from '../lib/Demo.svelte';
 	import Features from '../lib/Features.svelte';
 	let clientWidth;
+	const getDataSet = async () => {
+  const res = await fetch ("/api");//fetches the exported data from index.js. We are not sending a request body
+  const data = await res.json();//parses with json and assigns to data
+  if (res.ok){//if the response body is populated it will return it as props
+    return {
+    props: data
+    }
+  } else {
+    throw new Error(data);
+    }
+    }
+  let promise = getDataSet();
 </script>
 
 <!-- Loading app -->
@@ -32,11 +44,13 @@
 		<Landing />
 	</div>
 	<!-- Demo section -->
-	{#if clientWidth > 800}
+	<!--{#if clientWidth > 800}-->
 		<div class="DemoContainer">
-			<Demo />
+			{#await promise then output}
+				<Demo dataSet={output.props}/>
+			{/await}
 		</div>
-	{/if}
+	<!-- {/if} -->
 	<!-- Installation section -->
 	<div class="InstallationContainer">
 		<Installation />
